Unit Tests

+++ src/main/java/spring/security/security/service/LaptopService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import spring.security.security.dto.LaptopDto;
 import spring.security.security.mapper.LaptopMapper;
+import spring.security.security.model.Item;
 import spring.security.security.model.Laptop.Laptop;
 import spring.security.security.repository.CommentRepository;
 import spring.security.security.repository.ItemRepository;
@@ -40,9 +41,20 @@ public class LaptopService {
         return laptopMapper.toDto(laptopRepository.save(laptopMapper.toModel(laptopDto))) ;
     }
     public void deleteLaptop(Long id){
-        Long itemId = getLaptop(id).getItem().getId();
-        commentRepository.deleteByItem_Id(itemId);
-        laptopRepository.deleteById(id);
+//        Long itemId = getLaptop(id).getItem().getId();
+//        commentRepository.deleteByItem_Id(itemId);
+//        laptopRepository.deleteById(id);
+        LaptopDto laptopDto = getLaptop(id);
+        try {
+            Item item = laptopDto.getItem();
+            if (item != null) {
+                Long itemId = item.getId();
+                commentRepository.deleteByItem_Id(itemId);
+            }
+            laptopRepository.deleteById(id);
+        }catch (Exception e){
+            e.printStackTrace();
+        }
     }
 
 }
